refactor(client): collapse WRDistributionSVG passthrough into the only design that ships
WRDistributionSVG.tsx was a 30-line pure passthrough re-exporting
WRDistributionDesign2SVG.tsx since the original Design1 was deleted.
"Design2" is meaningless when there's only one design. Collapsed:

- git rm client/app/components/WRDistributionSVG.tsx (passthrough)
- git mv client/app/components/WRDistributionDesign2SVG.tsx
       → client/app/components/WRDistributionSVG.tsx
- Rename React component WRDistributionDesign2SVG → WRDistributionSVG
  and props interface WRDistributionDesign2Props → WRDistributionProps
  inside the moved file.

Public name imported by PlayerDetailInsightsTabs.tsx
(`WRDistributionSVG`) is unchanged — only the file holding the
implementation moved.

Tranche 3 of agents/runbooks/runbook-near-final-form-cleanup-2026-05-03.md.
Frontend release gate: lint clean, 95/95 tests, production build green.
